refactor: move temporal logic into TemporalInterlocutor (WIP) #2134
