+++ datashare-app/src/main/java/org/icij/datashare/mode/CommonMode.java
@@ -245,9 +245,9 @@ public abstract class CommonMode extends AbstractModule implements Closeable {
 
     @Provides @Singleton
     TaskManagerTemporal provideTaskManagerTemporal(
-            TemporalInterlocutor temporal, PropertiesProvider propertiesProvider) {
+            TemporalInterlocutor temporal, TaskRepository taskRepository, PropertiesProvider propertiesProvider) {
         return new TaskManagerTemporal(
-                temporal, Utils.getRoutingStrategy(propertiesProvider));
+                temporal, taskRepository, Utils.getRoutingStrategy(propertiesProvider));
     }
 
     @Provides @Singleton
@@ -417,7 +417,7 @@ public abstract class CommonMode extends AbstractModule implements Closeable {
                 workflows = TemporalHelper.discoverWorkflows(
                     "org.icij.datashare.tasks",
                     get(DatashareTaskFactory.class),
-                    temporal.client,
+                    temporal.getClient(),
                     Utils.getRoutingStrategy(propertiesProvider),
                     new Group(TaskGroupType.Java)
                 );
